refactor: extract validateBranchInArgs to solve.branch.lib.mjs (#1482)
Move validateBranchInArgs() from inline in telegram-bot.mjs to an
exported function in solve.branch.lib.mjs to keep telegram-bot.mjs
under the 1500-line CI limit. Tests now import the function directly.
